PythonBackend/requirements.txt
Flask
google-cloud-language
google-cloud-vision
tweepy
urllib
python-facebook-api
pyrebase